requirements.txt
kafka-python-ng==2.2.3
requests==2.31.0
fastapi==0.111.0
uvicorn==0.29.0
openai==1.30.1
python-dotenv==1.0.1
httpx==0.27.0
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.2
